test(tests/hoc/authentication.tsx): moved all tests out of authentication to their correct files
