+++ test/pages/index.test.tsx
@@ -12,7 +12,7 @@ import {
   mockRouter,
   renderWithTicket,
 } from '../test-utils';
-import { loadingAuth, notAuth } from '../hoc/authentication.test';
+import { loadingAuth, notAuth } from '../hoc/authentication';
 
 jest.mock('next/router', () => ({
   useRouter() {
